# trusted-pool-platform/api/openapi.yaml
openapi: 3.1.0
info:
  title: 可信有限成员订阅共享平台 Phase 2-A API
  version: 0.2.0
  description: |
    与当前 Go HTTP 实现一致的 Phase 2-A 持久运行时接口。Provision、Seat/Owner Assignment、Operation
    和 Provision credential claim 使用 PostgreSQL；数据库或 KMS 不可用时失败关闭。暂停、换员、结算、
    credential batch 和 control rotation evidence 尚未接入持久工作流，当前统一返回 503。
    风险窗口仍是可丢失的内存观察信号，不自动暂停。普通接口使用服务 Bearer API Key。
...
servers:
  - url: http://localhost:8092
security:
  - serviceBearer: []
paths:
...
  /api/v1/seats:
    post:
      operationId: provisionSeat
      summary: 在既有 Sub2API Group 中开通 Seat
      description: |
        平台先以 operation_id 幂等调用 Sub2API 创建不可交互 Principal、Subscription 和 API Key，
        仅在上游明确返回 ACTIVE 后创建本地 Seat。结果未知时返回 503，调用方必须以完全相同的请求体和
        operation_id 重放；成功响应不直接包含 credential。claim token 仅在首次明确成功响应中返回，
        平台只保存其 SHA-256 摘要，因此后续 provision 重放不会再次返回 token。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProvisionSeatRequest'}
      responses:
        '201':
          description: Sub2API 已明确完成开通，平台本地 Seat 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisionSeatEnvelope'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503':
          description: 上游不可用或结果未知；不得推断 Seat 已创建，须使用同一 operation_id 重放
  /api/v1/seats/{id}:
    get:
      operationId: getSeat
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Seat 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatEnvelope'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/settlements:
    get:
      operationId: listPendingSettlements
      summary: 列出阻断 Seat 冻结的待对账结算
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 100}
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/settlements/{settlement_id}:
    get:
      operationId: getPendingSettlement
      summary: 查询单条待对账结算
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/SettlementId'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/settlements/{settlement_id}/resolve:
    post:
      operationId: resolvePendingSettlement
      summary: 人工核验后解除单条结算屏障
      description: |
        Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。

        必须使用独立 TRUSTED_POOL_SETTLEMENT_API_KEY。平台把 operation_id 同时作为 Sub2API
        Idempotency-Key 提交；Sub2API 继续使用独立 settlement:resolve scope 校验集成客户端并写审计。
      security:
        - settlementBearer: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/SettlementId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResolvePendingSettlementRequest'}
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/actions/suspend:
    post:
      operationId: suspendSeat
      summary: 暂停并等待 Sub2API 排空
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503；后续实现仍须先排空至 FROZEN。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/actions/assign-temporary:
    post:
      operationId: assignTemporary
      summary: 冻结后分配临时成员
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/TargetUser'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/actions/restore:
    post:
      operationId: restoreOwner
      summary: 冻结后恢复正式成员
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/actions/replace:
    post:
      operationId: replacePermanently
      summary: 冻结后永久换员
      description: |
        Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。以下是后续阶段必须保留的安全门禁。

        必须提交本平台 control-rotation-evidence 接口签发的证据。Coordinator 会按 Seat Pool、当前
        Membership Epoch 和相邻新 Epoch 重新验证该证据所绑定的供应商变更证明引用及八个批次状态。
        Phase 1 不校验供应商签名或证明的密码学真实性；provider_attestation_ref 必须由运营或外部系统
        在签发证据前完成真实性核验。Sub2API 的 access_credential_rotated 仅代表 Seat 访问 Key 已轮换，
        不能冒充该证据。永久换员成功后，Pool 最小 Membership Epoch 提升到新 Epoch，旧 Epoch
        不能再 Seal 或 Activate 凭据批次。

        Phase 1 的该操作只验证暂停/冻结、Seat 访问 Key 轮换、外部 attestation 引用和控制凭据批次
        状态机；它不生成或签署 Recovery Root、Recovery Share 或 Manifest。因此 Phase 1 成功响应不构成
        完整永久交接，生产环境必须等待 Phase 2 recovery governance 门禁全部完成后才能确认永久换员。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/PermanentReplacement'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/operations/{id}:
    get:
      operationId: getOperation
      description: 只返回操作状态，不返回 credential_claim_token 或访问凭据。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200': {$ref: '#/components/responses/OperationResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/operations/{id}/reconcile:
    post:
      operationId: reconcileOperation
      description: Phase 2-A PostgreSQL Runtime 尚未开放显式 reconcile，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/operations/{id}/credential-ack:
    post:
      operationId: acknowledgeOperationCredential
      description: |
        目标成员使用原始开通或换员成功响应中一次性返回的随机 claim_token 领取 Seat 访问凭据。token 与目标成员
        和操作绑定且有短期 TTL。Provision 凭据返回前，平台必须先使用独立 `credential:ack` scope 向
        Sub2API 提交 credential SHA-256 fingerprint 并取得明确确认；确认结果未知时返回 503 且不披露凭据。
        成功后立即清除 token 摘要与凭据，重放不能再次领取。
        Sub2API ack 的 typed snapshot 会先持久化；Claim 尚未消费且 token 未过期时可在平台重启后继续领取。
        本接口最终采用 at-most-once 交付：平台将 Claim CAS 为 CLAIMED 并清密后，若对客户端的 HTTP 响应
        丢失，credential 永久不可恢复且不会二次披露，必须由管理员重新开通或轮换凭据。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CredentialClaimRequest'}
      responses:
        '200':
          description: 一次性交付访问凭据并完成 ack
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CredentialDeliveryEnvelope'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '410':
          description: claim token 已过期，待交付 credential 已销毁
        '502': {$ref: '#/components/responses/BadGateway'}
        '503':
          description: Sub2API ack 结果未知；失败关闭，不返回 credential，使用同一 token 重放本接口
  /api/v1/risk-windows:
    post:
      operationId: recordRiskObservation
      description: 原始指纹仅在请求处理内存中存在，聚合器只保留带密钥 HMAC；风险结果不触发暂停。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskObservation'
      responses:
        '202':
          description: 已聚合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskSummaryEnvelope'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/seats/{id}/risk-summary:
    get:
      operationId: listRiskSummaries
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: assignment_epoch
          in: query
          required: true
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        '200':
          description: 风险窗口列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskSummaryListEnvelope'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/credential-batches/{id}:
    get:
      operationId: getCredentialBatch
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      description: TRUSTED_POOL_API_KEY；仅供受控服务调用，不是最终用户会话或 RBAC
    settlementBearer:
      type: http
      scheme: bearer
      description: TRUSTED_POOL_SETTLEMENT_API_KEY；仅用于人工解除 pending settlement
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
  schemas:
...
    ProvisionSeatEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [seat, operation]
          properties:
            seat: {$ref: '#/components/schemas/Seat'}
            operation: {$ref: '#/components/schemas/Operation'}
...
  responses:
...
    BadRequest:
      description: 请求无效
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Unauthorized:
      description: Bearer 服务凭据缺失或无效
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    Conflict:
      description: Seat 状态、冻结快照或幂等 Key 冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    BadGateway:
      description: Sub2API 返回服务端错误或成功响应未通过一致性校验
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...